% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household}
\alias{household}
\title{New Zealand Census of Population and Dwellings household data}
\format{A data table with columns:
\describe{
  \item{geography}{Geographic level.  One of MB (meshblock), AU (area unit),
  Ward, TA (Territorial Authority), or RC (Regional Council).}
  \item{code}{Geography code.}
  \item{description}{Geography description.}
  \item{year}{Census year.  One of 2001, 2006, or 2013.}
  Other variables are documented in \code{\link{datadictionary}}.
}}
\source{
\url{http://www.stats.govt.nz/Census/2013-census/data-tables/meshblock-dataset.aspx}
}
\usage{
household
}
\description{
A dataset containing household counts at various geographic levels.
}
\keyword{datasets}

